# using Parent in integer rings instead of ParentWithGens