Let's try to get rid of the last direct inheritance to `ParentWithGens`.

(not quite the last one, see #37158 )

There remains many indirect inheritance, mostly through `Ring`, `Algebra`, `CommutativeRing` and their subclasses. This is a step towards cleaning and simplifying the old coercion framework.

The present changes is close to the heart of sage, so appropriate timings should be checked.

### :memo: Checklist

- [x] The title is concise, informative, and self-explanatory.
- [x] The description explains in detail what this PR is about.