@kwankyu 2024-01-29
> The present changes is close to the heart of sage, so appropriate timings should be checked.

What timings are necessary? Would you be explicit? Did you do the timings?