@tscrim 2024-01-31
I think coercions and conversions with common objects (finite fields, constant polynomials, rational numbers, etc.) would be good. Although I wouldn't expect really any changes in the timings...