@fchapoton 2024-01-31
I ran the tests of the `rings` folder and the time taken seemed not to change significantly. Not done scientifically..