@kwankyu 2024-01-31
I got, where m is a 1000*1000 matrix of integer entries,

Before this PR:
```
sage: m = load('m')
sage: timeit('m^3+m^2+m')
5 loops, best of 3: 1.14 s per loop
sage: timeit('1/3*m')
5 loops, best of 3: 447 ms per loop
```

After this PR:
```
sage: m = load('m')
sage: timeit('m^3+m^2+m')
5 loops, best of 3: 1.14 s per loop
sage: timeit('1/3*m')
5 loops, best of 3: 444 ms per loop
```

Perhaps not enough scientific either, but I don't perceive any performance regression.